Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -52,7 +52,7 @@ def get_average_age_for_season(data, season)
   data.each do |series, attributes|
     if series == season # identify correct season
       attributes.each do |key, value|
-        ages = ages + key["age"].to_i
+        ages = ages + key["age"].to_f
         persons = persons + 1
       end
     end
